+++ include/secondPassHelper.h
