+++ include/strHandle.h
