+++ src/core/assembler.c
