+++ src/core/strHandle.c
