+++ src/helpers/assemblerHelper.c
