+++ src/helpers/secondPassHelper.c
